moving interceptor and config to common
