+++ app/common/common-module.js
@@ -99,7 +99,8 @@ angular.module('ushahidi.common', [
         DELETE : 'delete'
     }
 })
-
+.factory('loading', require('./factories/loading.interceptor-factory.js'))
+.config(require('./configs/loading.interceptor-config.js'))
 .config(require('./configs/locale-config.js'))
 .run(require('./configs/ui-bootstrap-template-decorators.js'))
 .config(require('./configs/cache-config.js'))
