+++ app/common/configs/loading.interceptor-config.js
@@ -1,0 +1,3 @@
+module.exports = ['$httpProvider', function ($httpProvider) {
+        $httpProvider.interceptors.push('loading');
+    }];
